move action clip wrapper to gym wrappers
