+++ alf/algorithms/sac_algorithm.py
@@ -101,6 +101,7 @@ class SacAlgorithm(OffPolicyAlgorithm):
                  critic_optimizer=None,
                  alpha_optimizer=None,
                  gradient_clipping=None,
+                 clip_by_global_norm=False,
                  debug_summaries=False,
                  name="SacAlgorithm"):
         """Create a SacAlgorithm
@@ -134,6 +135,9 @@ class SacAlgorithm(OffPolicyAlgorithm):
             critic_optimizer (torch.optim.optimizer): The optimizer for critic.
             alpha_optimizer (torch.optim.optimizer): The optimizer for alpha.
             gradient_clipping (float): Norm length to clip gradients.
+            clip_by_global_norm (bool): If True, use `tensor_utils.clip_by_global_norm`
+                to clip gradient. If False, use `tensor_utils.clip_by_norms` for
+                each grad.
             debug_summaries (bool): True if debug summaries should be created.
             name (str): The name of this algorithm.
         """
@@ -158,6 +162,7 @@ class SacAlgorithm(OffPolicyAlgorithm):
             env=env,
             config=config,
             gradient_clipping=gradient_clipping,
+            clip_by_global_norm=clip_by_global_norm,
             debug_summaries=debug_summaries,
             name=name)
